remove edit and add landing page
